PBM-516 feature: PITR tests:  finished basic + dist transactions
