+++ e2e-tests/pkg/tests/sharded/test_pitr_trx.go
@@ -1,0 +1,5 @@
+package sharded
+
+func (c *Cluster) DistributedTrxPITR() {
+	c.DistributedTransactions(NewPitr(c))
+}
